# swagger.yml
openapi: 3.0.3
info:
  title: Podany API
  description: Private podcast RSS feed aggregator and web player API.
  version: 1.0.0
servers:
  - url: http://localhost:8788
    description: Local dev server
security:
  - cookieAuth: []
  - sessionTokenHeader: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: podcast_session
    sessionTokenHeader:
      type: apiKey
      in: header
      name: x-session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Download:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        episode_guid:
          type: string
        title:
          type: string
        audio_url:
          type: string
        subscription_id:
          type: integer
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /downloads:
    get:
      tags:
        - Downloads
      summary: List downloads for the current user
      responses:
        '200':
          description: Downloads
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Download'
    post:
      tags:
        - Downloads
      summary: Register a download for an episode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                episodeId:
                  type: string
                  description: Canonical episode identifier (downloads.id). Stored as the download's episode_guid anchor.
                title:
                  type: string
                audioUrl:
                  type: string
      responses:
        '200':
          description: Registered download
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  download:
                    $ref: '#/components/schemas/Download'
        '400':
          $ref: '#/components/responses/Error'
    delete:
      tags:
        - Downloads
      summary: Remove a download registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                episodeId:
                  type: string
                  description: Canonical episode identifier (downloads.id).
      responses:
        '200':
          description: Download removed
        '400':
          $ref: '#/components/responses/Error'
  /downloads/{id}/start:
    post:
      tags:
        - Downloads
      summary: Start a registered download
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Download started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  download:
                    $ref: '#/components/schemas/Download'
        '404':
          $ref: '#/components/responses/Error'
  /downloads/{id}:
    delete:
      tags:
        - Downloads
      summary: Delete a download record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Download deleted
        '404':
          $ref: '#/components/responses/Error'